requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

pandas

langchain
langchain-google-genai
langchain-experimental

python-jose[cryptography]

streamlit
st-aggrid
requests

httpx
